SwiftDemo - updated paths since restructure of repo + misc fixes
